refactor: editor InlineWidgetHelper.js
